repo_stream/requirements.txt
fastapi==0.124.2
uvicorn==0.38.0
aiofiles==25.1.0
pydantic==2.12.5
python-multipart==0.0.19
python-dotenv==1.0.1
